requirements.txt
streamlit>=1.33.0
openai>=1.0.0
opencv-python-headless>=4.5.5.64
pillow>=9.0.0